Scripts: enhance SCRIPT_COMMAND_TALK command to make Players talk (add 0x1 bit to datalong2 and command will expect Player instead of Creature for talking).
--HG--
branch : trunk

+++ src/server/game/Entities/Creature/Creature.cpp
@@ -2267,7 +2267,7 @@ void Creature::AllLootRemovedFromCorpse()
         else
             decayRate = 0.0f;
 
-        uint32 diff = (m_corpseRemoveTime - now) * decayRate;
+        uint32 diff = uint32((m_corpseRemoveTime - now) * decayRate);
 
         m_corpseRemoveTime -= diff;
         m_respawnTime -= diff;
